feat: chatbot — DTF sem blob, pausa auto 15min, cancelamento contextual, alerta de prazo
- DTF via WhatsApp cria sinalizador (triagem) sem blob; arquivos baixados manualmente pelo operador
- Confirmação WA disparada só quando operador avança para em_producao
- sync-outgoing pausa chatbot 15 min automaticamente após mensagem manual do operador
- Cancelamento contextual: pronto → alerta equipe, em_separacao → cancela + avisa, triagem/confirmando → cancela direto
- Pergunta de prazo (em_separacao/pronto) → needs_attention + resposta mínima, operador assume
- DTF em_producao + pergunta de status no idle → mesmo fluxo de alerta
- Aba Lifecycle no marketing com funil completo e toggle global lifecycle_ativo
- Settings: ALLOWED_KEYS atualizado, cron guarda steps 1-6 pelo lifecycle_ativo

+++ frontend/app/api/orders/[id]/route.ts
@@ -22,7 +22,7 @@ async function getOrder(id: string) {
             'productName',  i.product_name,
             'color',        i.color,
             'size',         i.size,
-            'qty',          i.qty,
+            'qty',          i.qty::int,
             'qtyConfirmed', i.qty_confirmed
           ) ORDER BY i.id
         ) FILTER (WHERE i.id IS NOT NULL),
